refactor: move business logic logic to  `helpers.py`
